Get provider name from og:site_name if available. Refs #79.
Issue #79 contains a list of potential sources for getting a 'provider name'.
However only og:site_name is the only reliable. The other options have slightly
different semantic meanings and therefore might return values that are not
suitable. Because of that this patch only adds support for og:site_name for now.